+++ src/TodoForm.jsx
@@ -1,0 +1,10 @@
+function TodoForm() {
+  return (
+    <>
+      <label htmlFor="todoTitle">Todo</label>
+      <input type="text" id="todoTitle" />
+      <button>Add Todo</button>
+    </>
+  );
+}
+export default TodoForm;
